separate creation and running event loop
